+++ NewPassword/functions/New-Password.ps1
@@ -18,6 +18,9 @@
     .PARAMETER UseUpperCase
         Switch to indicate to use upper case letters
 
+    .PARAMETER EncryptPassword
+        Switch to encrypt password
+
     .PARAMETER UseNumbers
         Switch to indicate to use numbers
 
@@ -36,13 +39,16 @@
     .EXAMPLE
         C:\> New-Password -Length 15 -UseSpecialChars -UseSpecialLetters -UseUpperCase -UserNumbers
 
+    .EXAMPLE
+        C:\> New-Password -Length 15 -UseSpecialChars -UseSpecialLetters -UseUpperCase -UserNumbers -EncryptPassword
+
     .NOTES
         None
     #>
 
     [cmdletbinding()]
     Param(
-        [Parameter(Mandatory = $True, ValueFromPipeline = $True, HelpMessage = "Indicate the word length between 1 and 25.")]
+        [Parameter(Mandatory = $True, ValueFromPipeline = $True, HelpMessage = "Indicate the password length between 1 and 25.")]
         [ValidateRange(1, 30)]
         [int]$Length,
         [switch]
@@ -54,7 +60,7 @@
         [switch]
         $UseNumbers,
         [switch]
-        $Logging
+        $EncryptPassword
     )
 
     begin {
@@ -67,13 +73,14 @@
     }
 
     process {
-        Write-Verbose "Checking switch parameters"
+        Write-PSFMessage -Level Verbose -Message "Checking switch parameters"
         if ($UseSpecialChars) { $password += $specialCharacters }
         if ($UseSpecialLetters) { $password += $specialLetters }
         if ($UseUpperCase) { $password += $upperCaseletters }
         if ($UseNumbers) { $password += $numbers }
         $password = -join ( $password | Get-Random -count $Length) # Default is just lower case letters
-
-        if($Logging) {Write-PSFMessage -Level Host -Message 'New Password is <c="yellow">{0}</c>' -StringValues $password}else {$password}
+        # Display the password without logging it
+        Write-Host "Password: $($password)"; Write-PSFMessage -Level Host -Message 'Password generated!'
+        if ($EncryptPassword) { New-EncryptedPassword -Password (ConvertTo-SecureString $password -AsPlainText -Force) }
     }
 }