+++ NewPassword/internal/tepp/assignment.ps1
@@ -2,3 +2,5 @@
 # Example:
 Register-PSFTeppArgumentCompleter -Command Get-Alcohol -Parameter Type -Name NewPassword.alcohol
 #>
+
+Register-PSFTeppArgumentCompleter -Command Get-PasswordParameters -Parameter Type -Name Get-PasswordParameters.settings